requirements.txt
streamlit
requests
matplotlib
fastapi
uvicorn[standard]
python-dotenv
requests
chromadb
sentence-transformers
torch
pdfplumber
pytesseract
pillow
